+++ src/main/java/io/github/astrapi69/swing/table/model/suffle/actions/RemoveAction.java
@@ -30,15 +30,16 @@ import javax.swing.*;
 
 import io.github.astrapi69.check.Check;
 import io.github.astrapi69.swing.GenericShuffleJTable;
+import io.github.astrapi69.swing.x.GenericShuffleJXTable;
 
 public class RemoveAction<T> extends AbstractAction
 {
 
 	private static final long serialVersionUID = 1L;
 
-	private final GenericShuffleJTable<T> shuffleJTable;
+	private final GenericShuffleJXTable<T> shuffleJTable;
 
-	public RemoveAction(final GenericShuffleJTable<T> shuffleJTable)
+	public RemoveAction(final GenericShuffleJXTable<T> shuffleJTable)
 	{
 		Check.get().notNull(shuffleJTable, "shuffleJTable");
 		this.shuffleJTable = shuffleJTable;